Style the header of the website and add an image
Use many div to structure the header. Use flex-direction column to
make it structurelike a real website. Add paddings and margins to all.
Use classes within classes with div to get the structure like wanted so
elements correctly split up where needed or align where needed as well.
Last add the button which was forgotten before. Add all the necessary
rules to make it as it should be like with border-radius to make
corners round