Add style to the body-box and set placeholder img
Align items at center and content as well  to follow the structure.
Add flex to make it all work accordingly. Put headline outside of flex
to put it above in the middle. Set div around the text to make it warp
in the next row. Set image placeholder for now with border