Add final style to the alert box and fix a bit
Set some divs to structure the alert box how i wanted to and with that
split the sections so the button is on the right side. Use margin-left
auto on the left box to keep space between always. Finally add the
footer styles and fix the image at the top a bit to make it not
shrink too much and make it bigger when possible. And thats it, the
first website or webpage i should say done